requirements.txt
requests>=2.25.0
pyyaml>=6.0
click>=8.0.0